[김홍주] Mission 3 수정
