+++ students/hongju/users/urls.py
@@ -2,6 +2,6 @@ from django.urls import path
 from users.views import SignUpView
 
 urlpatterns = [
-        path('signUp', SignUpView.as_view())
+        path('signup', SignUpView.as_view())
 ]
